Componentize treasury stats
